move `eudiKeyBindingStorage` to `HolderBindingKeyService`
